fix: retry_base_delay to float, pass model through ImageProvider (I3+I4)
Change PipelineConfig.retry_base_delay from int to float to allow
sub-second delays. Add model parameter to ImageProvider.generate()
Protocol and implementation so the model is read from config instead
of being hardcoded to "dall-e-3".

+++ src/story_video/models.py
@@ -323,7 +323,7 @@ class PipelineConfig(BaseModel):
 
     autonomous: bool = Field(default=False)
     max_retries: int = Field(default=3, ge=0)
-    retry_base_delay: int = Field(default=2, gt=0)
+    retry_base_delay: float = Field(default=2.0, gt=0)
     save_originals_on_revision: bool = Field(default=True)
